split UserListComponent from UserListScreen for testing, update tests
