+++ src/__tests__/App-test.js
@@ -1,14 +1,14 @@
-// /**
-//  * @format
-//  */
+/**
+ * @format
+ */
 
-// import 'react-native';
-// import React from 'react';
-// import App from '../App';
+import 'react-native';
+import React from 'react';
+import App from '../App';
 
-// // Note: test renderer must be required after react-native.
-// import renderer from 'react-test-renderer';
+// Note: test renderer must be required after react-native.
+import renderer from 'react-test-renderer';
 
-// it('renders correctly', () => {
-//   renderer.create(<App />);
-// });
+it('renders correctly', () => {
+  renderer.create(<App />);
+});
