+++ src/components/UserListScreen.js
@@ -1,31 +1,13 @@
 // @flow
 import * as React from 'react';
-import {
-  FlatList,
-  Button,
-  ActivityIndicator,
-  View,
-  StyleSheet,
-  Text,
-} from 'react-native';
+import {Button} from 'react-native';
+import UserListComponent from './UserListComponent';
 import type {RouteProp} from '@react-navigation/native';
 import type {StackNavigationProp} from '@react-navigation/stack';
 import type {RootState, RootStackParamList, UserDispatch} from '../types';
 import {useSelector, useDispatch} from 'react-redux';
-import UserItem from './UserItem';
 import {isFetchingUsersSelector, fetchUsersErrorSelector} from '../state';
 import {startFetchUsers} from '../UsersAction';
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-  },
-  errorText: {
-    textAlign: 'center',
-    fontSize: 16,
-  },
-});
 
 type UserListScreenProps = {
   navigation: StackNavigationProp<RootStackParamList, 'UserList'>,
@@ -53,32 +35,15 @@ const UserListScreen = ({
     });
   });
 
-  if (isFetching) {
-    return (
-      <View style={styles.container}>
-        <ActivityIndicator size="large" testID="loading" />
-      </View>
-    );
-  } else if (fetchUsersError) {
-    return (
-      <View style={styles.container}>
-        <Text style={styles.errorText}>
-          An error occurred: {fetchUsersError.message}
-        </Text>
-        <Button onPress={() => dispatch(startFetchUsers())} title="Retry" />
-      </View>
-    );
-  } else {
-    return (
-      <FlatList
-        data={users}
-        renderItem={({item: user}) => (
-          <UserItem user={user} navigation={navigation} />
-        )}
-        keyExtractor={user => user.id}
-      />
-    );
-  }
+  return (
+    <UserListComponent
+      isFetching={isFetching}
+      fetchUsersError={fetchUsersError}
+      users={users}
+      navigation={navigation}
+      dispatch={dispatch}
+    />
+  );
 };
 
 export default UserListScreen;
